change styling to tailwind
